Handle power loss during conversion
Reduce the likelyhood that dumps will be irrecoverably lost in case of
power loss or reset during conversion.

Only delete the dumps after all the conversion is done.

In case the dump files were deleted before the dump counter was
decremented, the counter will be out of sync with the actual number of
files under /dumps. If that happens, set the dumps count to 0 after all
the files have been processed.